Add a Global Attribute

Attach a global variable to the simulation.

add_global(.env, key, value)

Arguments

.env

the simulation environment.

key

the attribute name.

value

the value to set, either a numeric or a schedule, so that the global may change during the simulation.

Value

Returns the simulation environment.

See also