Get Process Parameters

Getters for processes (sources and arrivals) number of arrivals generated by a source, the name of the active arrival, an attribute from the active arrival or a global one, prioritization values, or the number of arrivals in an active batch.

Usage

get_n_generated(.env, sources)

get_trajectory(.env, sources)

get_name(.env)

get_start_time(.env)

get_attribute(.env, keys)

get_global(.env, keys)

get_prioritization(.env)

get_batch_size(.env)

Arguments

.env

the simulation environment.

sources

one or more resource names.

keys

the attribute name(s).

Details

get_n_generated returns the number of arrivals generated by the given sources. get_trajectory returns the trajectory to which they are attached (as a list).

get_name returns the number of the running arrival. get_start_time returns the start time of the running arrival. get_attribute returns a running arrival's attributes. If a provided key was not previously set, it returns a missing value. get_global returns a global attribute. get_prioritization returns a running arrival's prioritization values. get_name, get_start_time, get_attribute and get_prioritization are meant to be used inside a trajectory; otherwise, there will be no arrival running and these functions will throw an error.

See also