Set Source Parameters

Activities for modifying a source's trajectory or source object by name. Sources must be defined in the simulation environment (see add_generator, add_dataframe).

Usage

set_trajectory(.trj, sources, trajectory, ..., tag)

set_source(.trj, sources, object, ..., tag)

Arguments

.trj

the trajectory object.

sources

the name(s) of the source(s) or a function returning the name(s).

trajectory

the trajectory that the generated arrivals will follow.

...

unused.

tag

activity tag name to perform named rollbacks (see rollback) or just to better identify your activities.

object

a function modelling the interarrival times (if the source type is a generator; returning a negative value or a missing value stops the generator) or a data frame (if the source type is a data source).

Value

Returns the trajectory object.

See also

Examples

traj1 <- trajectory() %>%
  timeout(1)

traj2 <- trajectory() %>%
  set_source("dummy", function() 1) %>%
  set_trajectory("dummy", traj1) %>%
  timeout(2)

simmer() %>%
  add_generator("dummy", traj2, function() 2) %>%
  run(6) %>%
  get_mon_arrivals()
#>     name start_time end_time activity_time finished replication
#> 1 dummy0          2        4             2     TRUE           1
#> 2 dummy1          3        4             1     TRUE           1
#> 3 dummy2          4        5             1     TRUE           1