Clone/Synchronize Arrivals

Activities for defining a parallel fork and removing the copies. clone replicates an arrival n times (the original one + n-1 copies). synchronize removes all but one clone for each set of clones.

clone(.trj, n, ...)

synchronize(.trj, wait = TRUE, mon_all = FALSE)

Arguments

.trj
the trajectory object.

n
number of clones, accepts either a numeric or a callable object (a function) which must return a numeric.

...
a number of optional parallel sub-trajectories (or a list of sub-trajectories). Each clone will follow a different sub-trajectory if available.

wait
if FALSE, all clones but the first to arrive are removed. if TRUE (default), all clones but the last to arrive are removed.

mon_all
if TRUE, get_mon_arrivals will show one line per clone.

Value

Returns the trajectory object.