Run a Simulation

Execute steps until a given criterion.

run(.env, until = Inf, progress = NULL, steps = 10)

stepn(.env, n = 1)

Arguments

.env
the simulation environment.

until
stop time.

progress
optional callback to show the progress of the simulation. The completed ratio is periodically passed as argument to the callback.

steps
number of steps to show as progress (it takes effect only if progress is provided).

n
number of events to simulate.

Value

Returns the simulation environment.

See also