Set Resource Parameters

Activities for modifying a resource's server capacity or queue size, by name or a previously selected one.

set_capacity(.trj, resource, value, mod = c(NA, "+", "*"))

set_capacity_selected(.trj, value, id = 0, mod = c(NA, "+", "*"))

set_queue_size(.trj, resource, value, mod = c(NA, "+", "*"))

set_queue_size_selected(.trj, value, id = 0, mod = c(NA, "+", "*"))

Arguments

.trj
the trajectory object.

resource
the name of the resource.

value
new value to set.

mod
if set, values modify the attributes rather than substituting them.

id
selection identifier for nested usage.

Value

Returns the trajectory object.

See also