simmer: Discrete-Event Simulation for R

A process-oriented and trajectory-based Discrete-Event Simulation (DES) package for R. Designed to be a generic framework like SimPy or SimJulia, it leverages the power of Rcpp to boost the performance and turning DES in R feasible. As a noteworthy characteristic, simmer exploits the concept of trajectory: a common path in the simulation model for entities of the same type. It is pretty flexible and simple to use, and leverages the chaining/piping workflow introduced by the magrittr package.

References

Ucar I., Smeets B., Azcorra A. (2019). "simmer: Discrete-Event Simulation for R." Journal of Statistical Software, 90(2), 1-30. doi: 10.18637/jss.v090.i02 .

Ucar I., Hernández J.A., Serrano P., Azcorra A. (2018). "Design and Analysis of 5G Scenarios with simmer: An R Package for Fast DES Prototyping." IEEE Communications Magazine, 56(11), 145-151. doi: 10.1109/MCOM.2018.1700960 .

See also

Author

Iñaki Ucar and Bart Smeets

Examples

if (FALSE) {
# introduction to simmer
vignette("simmer-01-introduction")

# JSS paper available as vignette
vignette("simmer-02-jss")

# more vignettes
vignette(package = "simmer")
}